Fix direct navigation

+++ internal/pkg/filesystem/folder.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"log"
 	"os"
+	"sync"
 	"syscall"
 
 	"bazil.org/fuse"
@@ -17,10 +18,14 @@ import (
 
 type Folder struct {
 	fuse.Dirent
-	treeNode     *expanders.TreeNode
-	items        []*expanders.TreeNode
-	indexContent *expanders.ExpanderResponse
-	canDelete    bool
+	treeNode        *expanders.TreeNode
+	items           []*expanders.TreeNode
+	subFolders      []*Folder
+	indexContent    *expanders.ExpanderResponse
+	canDelete       bool
+	isParentDeleted func() bool
+	isBeingDeleted  bool
+	mu              sync.Mutex
 }
 
 var _ fs.Node = (*Folder)(nil)
@@ -37,6 +42,17 @@ func (d *Folder) Attr(ctx context.Context, a *fuse.Attr) error {
 var _ fs.NodeStringLookuper = (*Folder)(nil)
 
 func (d *Folder) Lookup(ctx context.Context, name string) (fs.Node, error) {
+	d.mu.Lock()
+	defer d.mu.Unlock()
+
+	if d.isDeleteInProgress() {
+		return nil, syscall.ENOENT
+	}
+
+	if d.indexContent == nil {
+		d.LoadNodeFromARM()
+	}
+
 	if name == indexFileName(d.treeNode, d.indexContent) {
 		file := &File{
 			treeNode: d.treeNode,
@@ -66,14 +82,31 @@ func (d *Folder) Lookup(ctx context.Context, name string) (fs.Node, error) {
 					Type: fuse.DT_Dir,
 					Name: nameFromTreeNode(treeNode),
 				},
+				isParentDeleted: func() bool {
+					d.mu.Lock()
+					defer d.mu.Unlock()
+
+					return d.isBeingDeleted
+				},
 			}
+			d.subFolders = append(d.subFolders, f)
 			return f, nil
 		}
 	}
+
+	log.Printf("Failed to match name: %s on folder %s", name, d.Name)
+
 	return nil, syscall.ENOENT
 }
 
 var _ fs.NodeRemover = (*Folder)(nil)
+
+func (d *Folder) isDeleteInProgress() bool {
+	if d.isBeingDeleted || d.isParentDeleted() {
+		return true
+	}
+	return false
+}
 
 // Todo currently `rm -r thing` gives an error deleting an RG but the delete is processed. Error:
 //
@@ -84,13 +117,30 @@ var _ fs.NodeRemover = (*Folder)(nil)
 // 30466/Deployments
 // rm: cannot remove '30466': Input/output error
 func (d *Folder) Remove(ctx context.Context, req *fuse.RemoveRequest) error {
-	for _, treeNode := range d.items {
-		if nameFromTreeNode(treeNode) == req.Name {
-			// if d.treeNode.ItemType == "subscription" {
-			// 	return fmt.Errorf("Can't delete subs, noop: %+v", req)
-			// }
+	d.mu.Lock()
+	defer d.mu.Unlock()
 
+	if d.isDeleteInProgress() {
+		return nil
+	}
+
+	for _, treeNode := range d.items {
+		if nameFromTreeNode(treeNode) == req.Name && treeNode.ItemType != "subResource" {
 			log.Printf("Found matching node '%s' doing delete: %+v", treeNode.Name+treeNode.ItemType, req)
+
+			// Mark folder as being deleted
+			for _, subFolder := range d.subFolders {
+				if subFolder.Name == req.Name {
+					func() {
+						subFolder.mu.Lock()
+						defer subFolder.mu.Unlock()
+						subFolder.isBeingDeleted = true
+					}()
+				}
+			}
+
+			log.Printf("---> Attemping delete on %s", req.Name)
+			// Start deletion
 			fallback := true
 			if treeNode.Expander != nil {
 				deleted, err := treeNode.Expander.Delete(ctx, treeNode)
@@ -99,7 +149,9 @@ func (d *Folder) Remove(ctx context.Context, req *fuse.RemoveRequest) error {
 			if fallback {
 				// fallback to ARM request to delete
 				_, err := armclient.LegacyInstance.DoRequest(ctx, "DELETE", treeNode.DeleteURL)
-				panic(err)
+				if err != nil {
+					panic(err)
+				}
 			}
 			log.Printf("Delete complete: %+v", req)
 		}
@@ -110,19 +162,30 @@ func (d *Folder) Remove(ctx context.Context, req *fuse.RemoveRequest) error {
 
 var _ fs.HandleReadDirAller = (*Folder)(nil)
 
+func (d *Folder) LoadNodeFromARM() {
+	rootContent, newItems, err := expanders.ExpandItem(ctx, d.treeNode)
+	if err != nil {
+		panic(err)
+	}
+
+	if *DemoMode {
+		rootContent.Response = views.StripSecretVals(rootContent.Response)
+	}
+
+	d.items = newItems
+	d.indexContent = rootContent
+}
+
 func (d *Folder) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
+	d.mu.Lock()
+	defer d.mu.Unlock()
+
+	if d.isDeleteInProgress() {
+		return nil, nil
+	}
+
 	if d.indexContent == nil {
-		rootContent, newItems, err := expanders.ExpandItem(ctx, d.treeNode)
-		if err != nil {
-			panic(err)
-		}
-
-		if *DemoMode {
-			rootContent.Response = views.StripSecretVals(rootContent.Response)
-		}
-
-		d.items = newItems
-		d.indexContent = rootContent
+		d.LoadNodeFromARM()
 	}
 
 	dirItems := make([]fuse.Dirent, len(d.items)+1)
